[3/N] Group together media-related code
