+++ tests/entrypoints/openai/test_sparse_tensor_validation.py
@@ -14,8 +14,7 @@ import pytest
 import torch
 
 from vllm.entrypoints.renderer import CompletionRenderer
-from vllm.multimodal.audio import AudioEmbeddingMediaIO
-from vllm.multimodal.image import ImageEmbeddingMediaIO
+from vllm.multimodal.media import AudioEmbeddingMediaIO, ImageEmbeddingMediaIO
 
 
 def _encode_tensor(tensor: torch.Tensor) -> bytes:
