+++ tests/entrypoints/openai/test_vision.py
@@ -8,7 +8,7 @@ import pytest
 import pytest_asyncio
 from transformers import AutoProcessor
 
-from vllm.multimodal.base import MediaWithBytes
+from vllm.multimodal.media import MediaWithBytes
 from vllm.multimodal.utils import encode_image_url, fetch_image
 from vllm.platforms import current_platform
